Updated Main hydrogen file
Added a file version. This will help keep track of what operands are
available and how the input file should be formatted (Not Applicable
yet).

Added data types. These data types are the same data types that are
used in Giraffe (since one of the major goals of Hydrogen is to run
Giraffe Bytecode)